backend/packages/telemetry/pyproject.toml
[project]
name = "selflytics-telemetry"
version = "0.1.0"
description = "Production-grade telemetry for Selflytics"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-logging>=0.41b0",
    "pydantic>=2.0.0",
    "google-cloud-logging>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telemetry"]
